apm-py/flask-auto-instr-extra/requirements.txt
backoff==1.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.42.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
opentelemetry-api==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-exporter-otlp-proto-grpc==1.7.1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-aws-lambda==0.26b1
opentelemetry-instrumentation-dbapi==0.26b1
opentelemetry-instrumentation-flask==0.26b1
opentelemetry-instrumentation-grpc==0.26b1
opentelemetry-instrumentation-logging==0.26b1
opentelemetry-instrumentation-sqlite3==0.26b1
opentelemetry-instrumentation-urllib==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-propagator-b3==1.7.1
opentelemetry-proto==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
opentelemetry-util-http==0.26b1
protobuf==3.19.1
requests==2.26.0
six==1.16.0
splunk-opentelemetry==1.2.0
thrift==0.15.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
